\docType{data}
\name{Actors}
\alias{Actors}
\title{Ages of Best Actor Oscar winners}
\format{A data frame with 32 observations on the following variable.

   \itemize{
    \item{\code{age}} {ages of Best Actor Oscar winners from 1970 to 2001}
  }}
\usage{
  data(Actors)
}
\description{
  The ages of Best Actor Oscar winners from 1970 to 2001
  when they received the award.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

