\docType{data}
\name{Autopremiums}
\alias{Autopremiums}
\title{Monthly Premiums of Auto Insurance}
\format{A data frame with 50 observations on the following variables.

   \itemize{
    \item{\code{Experience}} {driving experience (in years)}
    \item{\code{Female}} {0 = Male, 1 = Female}
    \item{\code{Premium}} {the monthly premium (in dollars)}
  }}
\usage{
  data(Autopremiums)
}
\description{
  The data were collected from a random sample of 50
  drivers insured with a certain company and having similar
  auto insurance coverage. For each driver the monthly
  premium was recorded along with other relevant
  information such as gender, age, driving experience,
  history of auto accidents, model and age of the car. The
  data set contains only a subset of the variables.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

