\docType{data}
\name{Baby}
\alias{Baby}
\title{Baby Cry Count and IQ}
\format{A data frame with 38 observations on the following variables.

   \itemize{
    \item{\code{cry.count}} {the number of distinct cry vocalizations within 20 seconds}
    \item{\code{IQ}} {IQ at 3 years of age}
  }}
\usage{
  data(Baby)
}
\description{
  A method for predicting IQ as soon as possible after
  birth could be important for early intervention in cases
  such as brain abnormalities or learning disabilities. It
  has been thought that greater infant vocalization (for
  instance, more crying) is associated with higher IQ. In
  1964, a study was undertaken to see if IQ at 3 years of
  age is associated with amount of crying at newborn age.
  In the study, 38 newborns were made to cry after being
  tapped on the foot, and the number of distinct cry
  vocalizations within 20 seconds was counted. The subjects
  were followed up at 3 years of age and their IQs were
  measured.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

