\docType{data}
\name{Birthweight}
\alias{Birthweight}
\title{Birth Weights of Newborns}
\format{A data frame with 125 observations on the following variables.

   \itemize{
    \item{\code{birthweight}}{weights of newborns (in grams)}
  }}
\usage{
  data(Birthweight)
}
\description{
  Some studies suggest that women having their first baby
  at age 35 or older are at increased risk of having a baby
  with a low birth weight. A medical researcher wanted to
  estimate the mean weight back in the population of
  newborns who are the first child for women over the age
  of 35. To this end, the researcher chose a random sample
  of 125 women age 35 or older who were pregnant with their
  first child and followed them through the pregnancy. The
  datafile linked below contains the birth weight (in
  grams) of the 125 newborns (women pregnant with more than
  one child were excluded from the study).
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

