\docType{data}
\name{Bodyimage}
\alias{Bodyimage}
\title{Body Image and Academic Performance of College Students}
\format{A data frame with 236 observations on the following variables.

   \itemize{
    \item{\code{Gender}} {sex \code{Female} \code{Male}}
    \item{\code{Height}} {self-reported heigh (in inches)}
    \item{\code{GPA}} {student's cumulative college GPA}
    \item{\code{HS_GPA}} {student's high school GPA for senior year}
    \item{\code{Seat}} {typical classroom sear location, \code{F} = Front, \code{M} = Middle, \code{B}}
    \item{\code{WtFeel}} {student's feeling about his/her weight: \code{UnderWt} \code{AboutRt} \code{OverWt}}
    \item{\code{Cheat}} {would they tell the instructor if he/she saw somebody cheating on exam? \code{Yes} \code{No}}
  }}
\usage{
  data(Bodyimage)
}
\description{
  A student survey was conducted at a major university.
  Data were collected from a random sample of 236
  undergraduate students, and the information that was
  collected included physical characteristics (such as
  height, handedness, etc.), study habits, academic
  performance and attitudes, and social behaviors.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

