\docType{data}
\name{Cellphones}
\alias{Cellphones}
\title{Cell Phones}
\format{A data frame with 312 observations on the following variables.

   \itemize{
    \item{\code{Math}} {SAT Math score}
    \item{\code{Verbal}} {SAT Verbal score}
    \item{\code{Credits}} {number of credits the student is registered for}
    \item{\code{Year}} {year in college (1 = First Year, 2 = Sophomore, 3 = Junior, 4 = Senior)}
    \item{\code{Exer}} {time (in minutes) spent exercising in a typical day}
    \item{\code{Sleep}} {time (in hours) spend sleeping in a typical day}
    \item{\code{Cell}} {do you own a cell phone? \code{yes} \code{no}}
    \item{\code{Veg}} {are you a vegetarian? \code{yes} \code{no} \code{some}}
  }}
\usage{
  data(Cellphones)
}
\description{
  College students at a large state university completed a
  survey about their academic and personal life. Questions
  ranged from "How many credits are you registered for this
  semester?" to "Would you define yourself as a
  vegetarian?"  Four sections of an introductory statistics
  course were chosen at random from all the sections of
  introductory statistics courses offered at the university
  in the semester when the survey was conducted, and the
  312 students who completed the survey were students
  registered in one of the four chosen sections.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

