% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collisions}
\alias{collisions}
\alias{select}
\alias{select.default}
\alias{select.ridgelm}
\alias{select.tbl}
\title{Avoiding inter-package collisions}
\usage{
select(object, ...)

\method{select}{tbl}(object, ...)

\method{select}{ridgelm}(object, ...)

\method{select}{default}(object, ...)
}
\arguments{
\item{object}{and R object}
}
\description{
Several popular packages export functions with the same names.  The \pkg{mosaic}
package attempts to help these other packge co-exist by creating generic functions
with the same name and dispatching to the appropriate function based on the (first)
arguemnt supplied.
}

