\docType{data}
\name{Computers}
\alias{Computers}
\title{Preferences for Computer Software}
\format{A data frame with 20783 observations on the following variables.
   \itemize{
    \item{\code{age}} {age of computer user (in years)}
    \item{\code{gender}} {a factor with levels \code{female} or \code{male}}
    \item{\code{comp}} {weekly time of computer use (in hours)}
 }}
\usage{
  data(Computers)
}
\description{
  A local internet service provider (ISP) created two new
  versions of its software, with alternative ways of
  implementing a new feature. To find the product that
  would lead to the highest satisfaction among customers,
  the ISP conducted an experiment comparing user preference
  for the two new versions versus the existing software.
  The ISP ideally wants to find out which of the three
  software products causes the highest user satisfaction.
  It has identified three major potential lurking variables
  that might affect user satisfaction: gender, age, and
  hours per week of computer use.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

