% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchData.R, R/fetchGapminder.R, R/fetchGoogle.R
\name{fetchData}
\alias{fetchData}
\alias{fetchGapminder}
\alias{fetchGapminder1}
\alias{fetchGoogle}
\title{Defunct functions now in the fetch package}
\usage{
fetchData(...)

fetchGapminder1(name, value.name = NULL)

fetchGapminder(..., all.cases = TRUE, all.vars = FALSE)

fetchGoogle(...)
}
\arguments{
\item{...}{arguments}

\item{name}{a character vector of length 1}

\item{value.name}{name of variable in resulting data frame}

\item{all.cases}{a logical indicating whether all cases should be included
or only complete cases}

\item{all.vars}{a logical, if \code{TRUE}, then \code{\dots} is ignored
and all available variables are fetched.}

\item{\dots}{character strings naming desired variables}
}
\value{
A data frame
}
\description{
These functions have been moved to the fetch package.

Fetch data originally obtained from Gapminder.
}

