% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMean.R, R/compareProportion.R,
%   R/deltaMethod.R, R/groupwiseModels.R, R/interval.R, R/mm.R, R/perctable.R,
%   R/xhistogram.R
\name{compareMean}
\alias{compareMean}
\alias{compareProportion}
\alias{deltaMethod}
\alias{gwm}
\alias{r.squared}
\alias{mm}
\alias{perctable}
\alias{proptable}
\alias{xhistogram}
\title{Defunct functions}
\usage{
compareMean(...)

compareProportion(...)

deltaMethod(...)

gwm(...)

r.squared(...)

mm(...)

perctable(...)

proptable(...)

xhistogram(...)
}
\arguments{
\item{\dots}{Additional arguments; currently ignored.}

\item{...}{arguments, ignored since the function is defunct}
}
\description{
The following functions were once a part of the mosaic pacakge but have
been removed.  In some cases, an alternative is available and is suggested
if you attempt to execute the function.
}
