% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaMethod.R
\name{deltaMethod}
\alias{deltaMethod}
\title{deltaMethod has moved to a separate package}
\usage{
deltaMethod(...)
}
\arguments{
\item{...}{arguments (ignored, since this is defunct)}
}
\description{
\code{deltaMethod()} has moved to its own separate package.
}
