% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diffmean}
\alias{diffmean}
\title{Difference in means}
\usage{
diffmean(...)
}
\arguments{
\item{...}{arguments passed to \code{mean}}
}
\description{
A wrapper around \code{diff(mean(...))} that facilitates better naming of the result
}
\examples{
diffmean( age ~ substance, data=HELPrct)
do(3) * diffmean(age ~ substance, data=HELPrct)
diffmean( age ~ sex, data=HELPrct)
do(3) * diffmean(age ~ sex, data=HELPrct)
}

