\docType{data}
\name{Drinking}
\alias{Drinking}
\title{Drinking Habits and Integrity of College Students}
\format{A data frame with 236 observations on the following variables.

   \itemize{
    \item{\code{Gender}} {sex \code{Female} \code{Male}}
    \item{\code{Alcohol}} {number of alcoholic beverage (standard units) consumed in a typical week}
    \item{\code{Height}} {self-reported height (in inches)}
    \item{\code{Cheat}} {would you tell the instructor if you saw somebody cheating on an exam, 0 = No, 1 = Yes}
  }}
\usage{
  data(Drinking)
}
\description{
  A student survey was conducted at a major university.
  Data were collected from a random sample of 236
  undergraduate students, and the information that was
  collected included physical characteristics (such as
  height and handedness), study habits, academic
  performance and attitudes, and social behaviors.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

