\docType{data}
\name{Drinks}
\alias{Drinks}
\title{Campus Drinking Rates}
\format{A data frame with 75 observations on the following variables.

   \itemize{
    \item{\code{drinks.per.week}}{number of alcoholic drinks (standard units) per week}
  }}
\usage{
  data(Drinks)
}
\description{
  A group of 75 college students were randomly sampled and
  asked about the number of alcoholic drinks they have in a
  typical week. The purpose of this study was to compare
  the drinking habits of the students at the college to the
  drinking habits of college students in general. In
  particular, the dean of students, who initiated this
  study, would like to check whether the mean number of
  alcoholic drinks that students at his college have in a
  typical week differs from the mean of U.S. college
  students in general, which is estimated to be 4.73.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

