\name{enterxtab}
\alias{enterxtab}
\title{Enter Cross Tables Interactively}
\usage{
  enterxtab()
}
\value{
  An \code{xtabs} object
}
\description{
  Simplified entering of cross tables from previously
  summarized data.
}
\author{
  Randall Pruim (\email{rpruim@calvin.edu})
}
\keyword{manipulate}

