\docType{data}
\name{EPA}
\alias{EPA}
\title{EPA Fuel Economy}
\format{A data frame with 71 observations on the following variables.

   \itemize{
    \item{\code{class}} {automobile type \code{TWO SEATERS}}
    \item{\code{mfr}} {automobile brand: one of \code{MERCEDES=BENZ} \code{PORSCHE} \code{BMW} \code{LAMBORGHINI} \code{CHRYSLER} \code{FERRARI} \code{ASTON MARTIN} \code{CADILLAC} \code{CHEVROLET} \code{HONDA} \code{LOTUS} \code{MAZDA} \code{NISSAN} \code{PONTIAC} \code{SATURN} \code{SPYKR}}
    \item{\code{line}} {model}
    \item{\code{displace}} {displacement of the engine}
    \item{\code{cylinders}} {number of cylinders}
    \item{\code{trans}} {transmission type}
    \item{\code{drive}} {drive system, 4 = , R = }
    \item{\code{city}} {city mileage (in miles per gallon)}
    \item{\code{hwy}} {highway mileage (in miles per gallon)}
    \item{\code{mpg}} {combined city and highway (in miles per gallon)}
  }}
\usage{
  data(EPA)
}
\description{
  Every year, the Environmental Protection Agency (EPA)
  collects data on fuel economy. With rising gasoline
  prices, consumers are using these figures as they decide
  which automobile to purchase. We will look at two-seater
  automobiles, many of which are sporty vehicles. The 2007
  EPA sample is used to test the hypothesis that the
  combined city and highway miles per gallon (mpg) of
  two-seater automobiles is greater than 20. The standard
  deviation for all vehicles is 4.7 mpg.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

