% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatingMeta.R
\name{f_flip}
\alias{f_flip}
\title{Flip lhs and rhs of formula}
\usage{
f_flip(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
a formula with the left and right hand sides reversed from the input.
}
\description{
Flip lhs and rhs of formula
}

