% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favstats.R
\name{fav_stats}
\alias{fav_stats}
\title{Some favorite statistical summaries}
\usage{
fav_stats(x, ..., na.rm = TRUE, type = 7)
}
\arguments{
\item{x}{numeric vector}

\item{...}{additional arguments (currently ignored)}

\item{na.rm}{boolean indicating whether missing data should be ignored}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile algorithms detailed
in the documentation for \code{\link[stats]{quantile}}}
}
\value{
A vector of statistical summaries
}
\description{
Likely you mean to be using \code{\link{favstats}}.  Each of these computes the 
mean, standard deviation, quartiles, sample size and number of missing values for a numeric vector,
but \code{\link{favstats}} can take a formula describing how these summary statistics 
should be aggregated across various subsets of the data.
}
\examples{
fav_stats(1:10)
fav_stats(faithful$eruptions)
favstats(Sepal.Length ~ Species, data=iris)  # Note: this is favstats() rather than fav_stats()
}
\keyword{stats}
