\name{.fetchData.storage.helper}
\alias{.fetchData.storage}
\alias{.fetchData.storage.helper}
\title{Internel fetch data function}
\usage{
  .fetchData.storage.helper()

  .fetchData.storage(library = FALSE, searchpath = FALSE,
    val = NULL, name = NULL, action)
}
\description{
  For internal use only
}
\keyword{internal}
\keyword{internel}
\keyword{mosaic}

