% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchGapminder1}
\alias{fetchGapminder}
\alias{fetchGapminder1}
\title{Fetch Gapminder data}
\usage{
fetchGapminder1(name, value.name = NULL)

fetchGapminder(..., all.cases = TRUE, all.vars = FALSE)
}
\arguments{
\item{name}{a character vector of length 1}

\item{value.name}{name of variable in resulting data frame}

\item{all.cases}{a logical indicating whether all cases should be included
or only complete cases}

\item{all.vars}{a logical, if \code{TRUE}, then \code{\dots} is ignored
and all available variables are fetched.}

\item{\dots}{character strings naming desired variables}
}
\value{
A data frame
}
\description{
Fetch data originally obtained from Gapminder.
}

