% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregatingAux.R
\name{formularise}
\alias{formularise}
\title{Convert lazy object into formulas}
\usage{
formularise(lazy_formula, envir = parent.frame())
}
\description{
Convert lazy object into formulas
}
\examples{
formularise(lazyeval::lazy(foo))
formularise(lazyeval::lazy(y ~ x))
bar <- a ~ b
formularise(lazyeval::lazy(bar))
}

