% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fractions}
\alias{fractions}
\title{Rational Approximation}
\usage{
fractions(x, cycles = 10, max.denominator = 2000, ...)
}
\arguments{
\item{x}{see \code{link[MASS]{fractions}} in the \pkg{MASS} package}

\item{cycles}{see \code{link[MASS]{fractions}} in the \pkg{MASS} package}

\item{max.denominator}{see \code{link[MASS]{fractions}} in the \pkg{MASS} package}

\item{...}{see \code{link[MASS]{fractions}} in the \pkg{MASS} package}
}
\description{
Rational Approximation from \pkg{MASS}.
}

