\docType{data}
\name{Friends}
\alias{Friends}
\title{Gender Concerns When Making Friends}
\format{A data frame with 1200 observations on the following variables.

   \itemize{
    \item{\code{easiest} }{answer to the question: "With whom do you find it easiest to make friends?" \code{No difference} \code{Opposite sex} or \code{Same sex}}
  }}
\usage{
  data(Friends)
}
\description{
  The survey is designed to investigate if there is any
  gender concern when students make friends. A total of
  1,200 U.S. college students were asked following
  question: "With whom do you find it easiest to make
  friends?" Possible responses were "No difference",
  "Opposite sex" and "Same sex".
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

