\docType{data}
\name{GestationLongevity}
\alias{GestationLongevity}
\title{Gestation Period and Longevity for Animals}
\format{A data frame with 40 observations on the following variables.

   \itemize{
    \item{\code{animal}} {name of the animal species}
    \item{\code{gestation}} {average gestation period of the species (in days)}
    \item{\code{longevity}} {average longevity of the species (in years)}
  }}
\usage{
  data(GestationLongevity)
}
\description{
  The average gestation period, or time of pregnancy, of an
  animal is closely related to its longevity (the length of
  its lifespan.) Data on the average gestation period and
  longevity (in captivity) of 40 different species of
  animals have been examined, with the purpose of examining
  how the gestation period of an animal is related to (or
  can be predicted from) its longevity.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

