\docType{data}
\name{Graduation}
\alias{Graduation}
\title{Graduation Rate}
\format{A data frame with 48 observations on the following variables.

   \itemize{
    \item{ontime}{percentage of students that graduated on time}
    \item{\code{college}}{a factor with levels \code{A} through \code{F}}
  }}
\usage{
  data(Graduation)
}
\description{
  The percentage of each entering Freshman class that
  graduated on time was recorded for each of six colleges
  at a major university over eight years.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

