\docType{data}
\name{Guns}
\alias{Guns}
\title{Ban on Assault Weapons}
\format{A data frame with 1200 observations on the following variables.

   \itemize{
    \item{\code{opinion}}{whether or not satisfied with the expiration of the ban on assault weapons: \code{satisfied} or \code{not satisfied}}
  }}
\usage{
  data(Guns)
}
\description{
  The U.S. federal ban on assault weapons expired in
  September 2004, which meant that after 10 years (since
  the ban was instituted in 1994) there were certain types
  of guns that could be manufactured legally again. A poll
  asked a random sample of 1,200 eligible voters (among
  other questions) whether they were satisfied with the
  fact that the law had expired. The data contains the
  results of this poll (Data were generated based on a poll
  conducted by NBC News/Wall Street Journal Poll).
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

