% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Heightweight}
\alias{Heightweight}
\title{Height and Weight}
\format{A data frame with 81 observations on the following variables.

 \itemize{
   \item{\code{female}} {0 = male, 1 = female}
   \item{\code{gender}} {a factor with levels \code{male} or \code{female}}
   \item{\code{height}} {subject height (in inches)}
   \item{\code{weight}} {subject weight (in pounds)}
 }}
\usage{
data(Heightweight)
}
\description{
The height and weight data collected from 57 males and 24 females for the
purpose of exploring how the weight of a person is related to his or her height.
}
\references{
Part of the Carnegie Mellon University Online Learning Initiative datasets.
}
\keyword{datasets}

