% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect objects}
\usage{
inspect(object, ...)
}
\description{
Print a short summary of the contents of an object.  Most useful as a way to
get a quick overview of the variables in data frame.
}
\examples{
inspect(Births78)
}

