% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evalFormula.R
\name{joinFrames}
\alias{joinFrames}
\alias{joinTwoFrames}
\title{Join data frames}
\usage{
joinFrames(...)

joinTwoFrames(left, right)
}
\arguments{
\item{left,right}{data frames}

\item{\dots}{data frames to be joined}
}
\value{
a data frame containing columns from each of data frames being joined.
}
\description{
Join data frames
}

