% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MAD}
\alias{MAD}
\alias{SAD}
\title{All pairs mean and sum of absolute differences}
\usage{
MAD(x, ..., data, groups = NULL, ..fun.. = MAD_,
  na.rm = getOption("na.rm", FALSE))

SAD(x, ..., data, groups = NULL, ..fun.. = SAD_,
  na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{a numeric vector or a formula.}

\item{...}{additional arguments passed through to \code{MAD_}
or \code{SAD_}.  If \code{x} is a formala, \code{...} should
include an argument named \code{data} if the intent is to
interpret the formala in a data frame.}

\item{na.rm}{a logical indicating whether NAs should be removed before
calculaing.}
}
\value{
the mean or sum of the absolute differences between each pair
of values in \code{c(x,...)}.
}
\description{
All pairs mean and sum of absolute differences
}
\examples{
SAD(1:3)
MAD(1:3)
MAD(~eruptions, data=faithful)
}
\seealso{
\code{link{mad}}, \code{\link{MAD_}}
}

