% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MAD_}
\alias{MAD_}
\alias{SAD_}
\title{All pairs mean and sum of absolute differences}
\usage{
MAD_(x, ..., na.rm = getOption("na.omit", FALSE))

SAD_(x, ..., na.rm = getOption("na.omit", FALSE))
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{if present, appended to x}

\item{na.rm}{a logical indicating whether NAs should be removed before
calculaing.}
}
\value{
the mean or sum of the absolute differences between each pair
of values in \code{c(x,...)}.
}
\description{
All pairs mean and sum of absolute differences
}
\seealso{
\code{link{mad}}
}

