% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Defunct modeling function}
\usage{
mm(formula, data = parent.frame(), fun = mean, drop = TRUE, ...)
}
\arguments{
\item{formula}{A formula.  The left-hand side specifies the variable over
which the mean will be taken.  The right-hand side gives the grouping variables,
separated by \code{&}.}

\item{data}{A data frame to which the formula variables refer.  If not specified, variables
will be taken from the current environment.}

\item{fun}{The function used to calculate the means.  Default: \code{mean}.}

\item{drop}{Logical flag indicating whether to drop unoccupied groups.  Default \code{TRUE}.
NOT YET IMPLEMENTED.}

\item{\dots}{Additional arguments to be passed to the \code{fun} doing the calculation.}
}
\description{
The function has been removed from \pkg{mosaic}.
}
