% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatingAux.R, R/do.R, R/integrateODE.R,
%   R/symbolicInt.R, R/utils.R
\name{.is.formula}
\alias{.is.formula}
\alias{.is.simple.formula}
\alias{.simple.part}
\alias{.make.data.frame}
\alias{.merge_data_frames}
\alias{.squash_names}
\alias{fetchDynamics}
\alias{dynamicsFunction}
\alias{rkFunction}
\alias{.makeNice}
\alias{.do.safe.call}
\title{Check if formula}
\usage{
.is.formula(x)

.is.simple.formula(x)

.simple.part(x)

.make.data.frame(x)

.merge_data_frames(a, b)

.squash_names(object, sep = ":")

fetchDynamics(x)

dynamicsFunction(DE, additionalAssignments = list())

rkFunction(DE, additionalArguments = list())

.makeNice(form, params = all.vars(form))

.do.safe.call(what, args, quote = FALSE, envir = parent.frame(), ...)
}
\arguments{
\item{x}{an object}

\item{a}{a data frame}

\item{b}{a data frame}

\item{object}{an object}

\item{sep}{a character}

\item{DE}{representation of DE, the result of fetchDynamics}

\item{additionalAssignments, }{a list
return a function}

\item{what}{either a function or a non-empty character string naming the function to be called.}

\item{args}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{quote}{a logical value indicating whether to quote the arguments.}

\item{envir}{an environment within which to evaluate the call. 
This will be most useful if what is a character string and the arguments are symbols or quoted expressions.}

\item{x}{a formula}

\item{x}{a formula}

\item{x}{object to be converted}

\item{x}{a list}

\item{DE}{representation of DE, the result of fetchDynamics}

\item{additionalAssignments, }{a list
return a function}
}
\value{
TRUE for a formula, FALSE otherwise, even if evaluation throws an error

TRUE if formula has no left-hand side or a simple right-hand side 
(e.g., \code{NULL}, ., 1,  or 0)

simple part of formula or NULL if formula is not simple

a data frame

a data frame

a character vector

a list with two slots: names and functions

The result of the (evaluated) function call.
}
\description{
Check if formula

Check for simple formula

Extract simple part from formula

parameters are stored as extra arguments
the order of the dynamical variables (and "t") is important and will be used
later

Create a functions with a vector argument of state, for use in rk()

Simplifying expressions, e.g. pure numbers go to numbers
Written by Aaron Mayerson, May 2013
}
\details{
\code{.make.data.frame} converts things to a data frame

\code{.merge_data_frames} is a wrapper around merge

\code{.squash_names} squashes names of a data frame into a single string

\code{.do.safe.call} avoids conflicts between named arguments and ... by taking named arguments 
preferentially.
}
\seealso{
\code{\link{do.call}}
}
\keyword{internal}
