% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic}
\alias{mosaic}
\title{Create a Mosaic Widget}
\usage{
mosaic(
  spec,
  ...,
  api = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  baseURL = NULL
)
}
\arguments{
\item{spec}{A list representing the Mosaic specification. See the
\href{https://idl.uw.edu/mosaic/api/spec/format.html}{specification format reference}
for details.}

\item{...}{Named data frames to be used in the plot. Each data frame will be
added to the spec in an efficient format so it can be used in the plot.
Can only be used in non-Shiny contexts or when not using the \code{api} parameter.
If using the \code{api}, insert the data directly into the duckdb connection.}

\item{api}{An optional string identifier for a server-side data API. This can be
used to fetch data from a server-side DuckDB database. It's only used in Shiny apps
in combination with \code{\link[=mosaicServer]{mosaicServer()}}.}

\item{width}{The width of the widget (optional). Note that this overrides any
width specified in the spec.}

\item{height}{The height of the widget (optional). Note that this overrides any
height specified in the spec.}

\item{elementId}{An optional element ID for the widget.}

\item{baseURL}{The baseURL used to acquire files. See additional docs in
\href{https://idl.uw.edu/mosaic/api/spec/parser-generators.html#asttodom}{astToDom}.}
}
\description{
Create a Mosaic widget that can be used in R Markdown documents,
Shiny apps, and the R console.
}
