% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaicServer}
\alias{mosaicServer}
\title{Mosaic Server}
\usage{
mosaicServer(id, connection)
}
\arguments{
\item{id}{The id of the module.}

\item{connection}{A DBI connection to a DuckDB database.}
}
\value{
A shiny reactive that returns the id of the API.
}
\description{
Creates and registers a global mosaic API, coordinator and connector
that can be acessed from js using:
}
\details{
\if{html}{\out{<div class="sourceCode js">}}\preformatted{window.[id]
window.[id + "_coordinator"]
window.[id + "_connector"]
}\if{html}{\out{</div>}}

This function is intended to be used in Shiny apps in combination with
the \code{api} parameter of the \code{\link[=mosaic]{mosaic()}} function. Allowing you to create
plots that fetch data from a server-side DuckDB database, so you don't
need to send large datasets to the client.
}
