% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/groupwiseModels.R
\name{MSPE}
\alias{MSPE}
\title{Mean Squared Prediction Error}
\usage{
MSPE(mod, test_data)
}
\description{
Mean Squared Prediction Error
}
\examples{
HELP <- HELPrct \%>\% sample_frac(.3)
MSPE( gwm( age ~ sex, data = HELP), HELPrct)
MSPE( gwm( age ~ 1, data = HELP), HELPrct)
MSPE( gwm( age ~ sex + homeless, data = HELP), HELPrct)
MSPE( gwm( sex ~ 1, data = HELP), HELPrct)
MSPE( gwm( sex ~ homeless, data = HELP), HELPrct)
MSPE( gwm( sex ~ homeless + substance, data = HELP), HELPrct)
}

