% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.msummary.lm}
\alias{msummary}
\alias{msummary.default}
\alias{msummary.glm}
\alias{msummary.lm}
\alias{print.msummary.glm}
\alias{print.msummary.lm}
\title{Modified summaries}
\usage{
\method{print}{msummary.lm}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)

\method{print}{msummary.glm}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)

msummary(object, ...)

\method{msummary}{default}(object, ...)

\method{msummary}{lm}(object, ...)

\method{msummary}{glm}(object, ...)
}
\description{
\code{msummary} provides modified summary objects that typically produce
output that is either identical to or somewhat terser than their
\code{\link{summary}} analogs.  The contents of the object itself are unchanged
(except for an augmented class) so that other downstream functions should work as
before.
}

