% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mustangs.R
\docType{data}
\name{Mustangs}
\alias{Mustangs}
\title{Mustang Prices}
\format{
A data frame with 25 observations on the following 3 variables.
\itemize{
\item{\code{Age }}{age of vehicle in years}
\item{\code{Miles }}{1000s of miles driven}
\item{\code{Price }}{selling price in 1000s USD}
}
}
\source{
These data were used in a "resampling bake-off" hosted by Robin Lock.
}
\usage{
data(Mustangs)
}
\description{
Mustang Prices
}
\details{
#' @docType data

A student collected data on the selling prices for a sample of used Mustang
cars being offered for sale at an internet website.
}
\keyword{datasets}
