% Generated by roxygen2 (4.0.2): do not edit by hand
\name{named}
\alias{named}
\alias{named_among}
\alias{unnamed}
\title{List extraction}
\usage{
named(l)

unnamed(l)

named_among(l, n)
}
\arguments{
\item{l}{a list}

\item{n}{a vector of character strings (potential names)}
}
\value{
a sublist of \code{l} determined by \code{names(l)}
}
\description{
These functions create subsets of lists based on their names
}

