% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{NHANES}
\alias{NHANES}
\title{Body Shape and related measurements from NHANES 1999-2004}
\format{A data frame with 31126 observations on the following 53 variables.
\itemize{
\item{\code{seqn} matching code for NHANES mortality data}
\item{\code{sex} \code{"male"} or \code{"female"}}
\item{\code{age} in years}
\item{\code{pregnant} \code{"yes"} or \code{"no"}}
\item{\code{ethnicity} Mexican American,
Other Hispanic, Non-Hispanic White, Non-Hispanic Black, or Other/Multi
}
\item{\code{death} \code{"alive"},  \code{"cardiovascular death"} or \code{"other death"}}
\item{\code{followup} months of follow up (for mortality data)}
\item{\code{smoker}
\code{"yes"} or \code{"no"}}
\item{\code{diabetic}
\code{"yes"} or \code{"no"}}
\item{\code{height} height (meters)}
\item{\code{weight} weight (kilograms)}
\item{\code{waist} waist circumference (meters)}
\item{\code{wci} the proposed body shape index}
\item{\code{bmi} body mass index}
\item{\code{ptfat} percent trunk fat}
\item{\code{tfat} mass of trunk fat}
\item{\code{lfat} limb fat}
\item{\code{llean} limb lean tissue}
\item{\code{lean_bmi} lean-tissue only BMI}
\item{\code{fat_bmi} fat-only BMI}
\item{\code{bone_bmi} bone BMI}
\item{\code{pct_fat} percent fat}
\item{\code{bmd} bone mineral density}
\item{\code{fmhm_other} Framingham risk score}
\item{\code{hdl} HDL cholesterol}
\item{\code{chol} cholesterol (LDL?)}
\item{\code{bps} systolic blood pressure, mmHg}
\item{\code{bpd} diastolic blood pressure, mmHg}
\item{\code{income} ratio of family income to poverty threshold. 5 is ratio greater than or equal to 5}
\item{\code{zheight} z-score stratified by age and sex}
\item{\code{zweight} z-score stratified by age and sex}
\item{\code{zwaist} z-score stratified by age and sex}
\item{\code{zwci} z-score stratified by age and sex}
\item{\code{zbmi} z-score stratified by age and sex}
\item{\code{zptfat} z-score stratified by age and sex}
\item{\code{ztfat} z-score stratified by age and sex}
\item{\code{zlfat} z-score stratified by age and sex}
\item{\code{zllean} z-score stratified by age and sex}
\item{\code{zlbmi} z-score stratified by age and sex}
\item{\code{zfbmi} z-score stratified by age and sex}
\item{\code{zbbmi} z-score stratified by age and sex}
\item{\code{zpf} z-score stratified by age and sex}
\item{\code{zbmd} z-score stratified by age and sex}
\item{\code{zfmhm_other} z-score stratified by age and sex}
\item{\code{zhdl} z-score stratified by age and sex}
\item{\code{zchol} z-score stratified by age and sex}
\item{\code{zbps} z-score stratified by age and sex}
\item{\code{zbpd} z-score stratified by age and sex}
\item{\code{zincome} z-score stratified by age and sex}
\item{\code{pop_weight} population weight, for adding up over the US population }
\item{\code{psu} primary sampling unit}
\item{\code{stratum} sampling stratum}
\item{\code{zwh} z-score of waist/height ratio}
}}
\source{
These data were assempled from NHANES 1999-2004 sources by Dr. Nir
Krakauer at City College of New York.
}
\description{
These data were collected as part of the NHANES 1999-2004 study.
}
\examples{
data(NHANES)
}
\references{
Nir Y. Krakauer, Jesse C. Krakauer, "A New Body Shape Index
Predicts Mortality Hazard Independently of Body Mass Index" PloS one, 7 (7)
PMID: 22815707, 2012.
\url{http://www.plosone.org/article/info:doi/10.1371/journal.pone.0039504}
}
\keyword{datasets}

