\docType{data}
\name{Nightlight}
\alias{Nightlight}
\title{Night Lights and Nearsightedness}
\format{A data frame with 479 observations on the following variables.

   \itemize{
    \item{\code{Light}} {lighting conditions \code{no light} \code{night light} \code{lamp}}
    \item{\code{Nearsightedness}} {whether or not the subjects later became nearsighted \code{Yes} \code{No}}
  }}
\usage{
  data(Nightlight)
}
\description{
  An Associated Press article captured the attention of
  readers with the headline "Night lights bad for kids?"
  The article was based on a 1999 study at the University
  of Pennsylvania and Children's Hospital of Philadelphia,
  in which parents were surveyed about the lighting
  conditions under which their children slept between birth
  and age 2 (lamp, night-light, or no light) and whether or
  not their children developed nearsightedness (myopia).
  The purpose of the study was to explore the effect of a
  young child's nighttime exposure to light on later
  nearsightedness.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

