% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ntiles}
\alias{ntiles}
\title{Create vector based on roughly equally sized groups}
\usage{
ntiles(x, n = 3, format = c("rank", "interval", "center", "left", "right"),
  digits = 3)
}
\arguments{
  \item{x}{a numeric vector}

  \item{n}{(approximate) number of quantiles}

  \item{format}{a specification of desired output format.}

  \item{digits}{desired number of digits for labeling of
  factors.}
}
\value{
a vector.  The type of vector will depend on \code{format}.
}
\description{
Create vector based on roughly equally sized groups
}
\examples{
tally( ~ ntiles(1:50, 4) )
tally( ~ ntiles(1:50, 4, format="center") )
tally( ~ ntiles(1:50, 4, format="interval") )
tally( ~ ntiles(1:50, 4, format="left") )
tally( ~ ntiles(1:50, 4, format="right") )
}

