% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oddsRatio}
\alias{oddsRatio}
\alias{print.oddsRatio}
\alias{relrisk}
\title{Odds Ratio and Relative Risk for 2 x 2 Contingency Tables}
\usage{
oddsRatio(x, conf.level = 0.95, verbose = !quiet, quiet = TRUE,
  digits = 3)

\method{print}{oddsRatio}(x, digits = 4, ...)

relrisk(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a 2 X 2 matrix, data frame or table of counts}

\item{conf.level}{the confidence interval level}

\item{digits}{number of digits to display}

\item{\dots}{additional arguments}
}
\value{
an odds ratio or relative risk.  If \code{verpose} is true,
more details and the confidence intervals are displayed.
}
\description{
This function calculates the odds ratio and relative risk for a 2 x 2
contingency table and a
confidence interval (default \code{conf.level} is 95 percent) for the
each estimate. \code{x} should be a matrix, data frame or table. "Successes"
should be located in column 1 of \code{x}, and the treatment of interest
should be located in row 2. The odds ratio is calculated as (Odds row 2) /
(Odds row 1). The confidence interval is calculated from the log(OR) and
backtransformed.
}
\examples{
M1 <- matrix(c(14, 38, 51, 11), nrow = 2)
M1
oddsRatio(M1)

M2 <- matrix(c(18515, 18496, 1427, 1438), nrow = 2)
rownames(M2) <- c("Placebo", "Aspirin")
colnames(M2) <- c("No", "Yes")
M2
oddsRatio(M2)
oddsRatio(M2, verbose=TRUE)
relrisk(M2, verbose=TRUE)
relrisk(tally(~ homeless + sex, data=HELPrct) )
do(3) * relrisk( tally( ~ homeless + shuffle(sex), data=HELPrct) )
}
\author{
Kevin Middleton (\email{kmm@csusb.edu}); modified by
R Pruim.
}
\seealso{
\code{\link{chisq.test}}, \code{\link{fisher.test}}
}
\keyword{stats}

