\docType{data}
\name{Olympics}
\alias{Olympics}
\title{Winning Times for 1,500 Meter Race}
\format{A data frame with 24 observations on the following variables.
   \itemize{
    \item{\code{Year}} {the year of the Olympic Games, from 1896 to 2000}
    \item{\code{Time}} {the winning time for the 1,500 meter race (in seconds)}
  }}
\usage{
  data(Olympics)
}
\description{
  The modern Olympic Games have changed dramatically since
  their inception in 1896. For example, many commentators
  have remarked on the change in the quality of athletic
  performances from year to year. Regression will allow us
  to investigate the change in winning times for one event:
  the 1,500 meter race.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

