\name{plot.mosaic}
\alias{plot.mosaic}
\title{
Plot a Mosaic Graph
}
\description{
The \emph{plot} method portrays mosaic graphs or minimum spanning trees (MSTs), with vertices representing the original data set. Plots can be output directly to a pdf file, and they can be colored or enhanced by including rectangles that indicate the contribution of each edge to the overall area estimate.
}
\usage{
plot.mosaic(m, add = F, pdf = F, mar = 0, xmin = NA, xmax = NA, ymin = NA,
ymax = NA, cex = 1, lines = T, lwd = 1, squares = F, fill = T, outline = T,
col = NA, h = 0.5, s = 0, v = 0.5, alpha = 0.5, pts.black = T,
lines.black = T)
}
\arguments{
  \item{m}{
A mosaic object generated by the \link{mgraph} function or \link{tgraph} function (required).
}
  \item{add}{
Create a new plot (F) or add to an existing plot (T)?
}
  \item{pdf}{
Should a pdf version of the plot be printed to the working directory? Ignored if new = F.
}
  \item{mar}{
The margin around the plot area.
}
  \item{xmin}{
The x-axis minimum.
}
  \item{xmax}{
The x-axis maximum.
}
  \item{ymin}{
The y-axis minimum.
}
  \item{ymax}{
The y-axis maximum.
}
  \item{cex}{
The point size magnification.
}
  \item{lines}{
Should lines be printed to represent the connected edges?
}
  \item{lwd}{
The line thickness.
}
  \item{squares}{
Should squares be printed to represent the area estimates?.
}
  \item{fill}{
Should the outline of the mosaic be filled with the designated color?
}
  \item{outline}{
Should the outline of the mosaic be drawn in the designated color? Has no effect if \emph{lines} is true, and only matters when \emph{fill} is true if there are isolated mosaic pieces connected by bridges.
}
  \item{col}{
The colour. The accepted values are red, orange, yellow, green, lightblue, blue, purple, and pink. The value must be placed in quotes. Overrides h, s, and v.
}
  \item{h}{
The hue, ranging from red (0) to pink (1).
}
  \item{s}{
The saturation, ranging from greyscale (0) to fully saturated (1).
}
  \item{v}{
The value (brightness)s, ranging from dark (0) to light (1).
}
  \item{alpha}{
The transparency, ranging from invisible (0) to solid (1).
}
  \item{pts.black}{
Should the points be black (T) or the same color as everything else (F)?
}
  \item{lines.black}{
Should the lines be black (T) or the same color as everything else (F)?
}
}
\details{
The mosaic object generated by \link{mgraph} is passed automatically to \link{mhull} if \emph{fill} is true or \emph{outline} is true.

If an MST is being drawn, lines is set to T and squares, fill, and outline are all set to F automatically.
}
\references{
Alroy, J. A simple graph theoretic method provides accurate range area estimates. To be submitted to \emph{Ecology Letters}.
}
\author{
John Alroy
}
\examples{
# define a mosaic object
m <- mgraph(runif(50),runif(50))

# make the data points light purple
plot(m, h = 0.7, s = 0.5, v = 1, pts.black = F)

# plot squares only, omitting edges
plot(m, cex = 0, lines = F, squares = T)

# plot pink squares
plot(m, h = 0.95, s = 1, v = 1, squares = T)

# only plot the area filled by the mosaic, with no points or lines
plot(m, cex = 0, lines = F, fill = T, outline = T, h = 0.15, s = 1, v = 1)

# define an MST instead of a mosaic
t <- tgraph(runif(50),runif(50))

# make both the lines and the points blue
plot(t, h = 0.6, s = 1, v = 1, pts.black = F, lines.black = F)

# plot a mosaic on top of an MST, keeping the points black
x <- runif(50)
y <- runif(50)
m <- mgraph(x,y)
t <- tgraph(x,y)
plot(t, lwd = 3, h = 0.6, s = 1, v = 1, lines.black = F)
plot(m, add = T, h = 0.05, s = 1, v = 1, lines.black = F)
}
