\docType{data}
\name{Population}
\alias{Population}
\title{Population Taking Introductory Statistics Courses}
\format{A data frame with 1129 observations on the following variables.

   \itemize{
    \item{\code{Course}} {purpose of taking introductory statistics courses \code{natural science} \code{social science} \code{business}}
    \item{\code{Handed}} {handedness \code{right-handed} \code{left-handed}}
    \item{\code{Sex}} {a factor with levels \code{female} or \code{male}}
    \item{\code{Verbal}} {SAT verbal scores up to 800}
    \item{\code{Age}} {age (in years)}
  }}
\usage{
  data(Population)
}
\description{
  1,129 students took introductory statistics courses for
  different purposes: business, social sciences, or natural
  sciences. The researcher were also interested in the
  values of four specific variables for the population:
  handedness (right-handed or left-handed), sex, SAT Verbal
  score, and age.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

