\docType{data}
\name{Pregnancy}
\alias{Pregnancy}
\title{Length of Pregnancy}
\format{A data frame with 25 observations on the following variables.

   \itemize{
    \item{\code{length} }{the length of pregnancy (in days) of women who were smoking and/or drinking alcohol during their pregnancy}
  }}
\usage{
  data(Pregnancy)
}
\description{
  The length of human pregnancy is known to have a mean of
  266 days and a standard deviation of 16 days. Based on
  records from a large women's hospital, a random sample of
  25 women who were smoking and/or drinking alcohol during
  their pregnancy and their pregnancy lengths are recorded.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

