% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_summary_lm}
\alias{print_summary_glm}
\alias{print_summary_lm}
\title{Alternative print functions}
\usage{
print_summary_lm(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)

print_summary_glm(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\description{
The functions provide alternatives to \code{print} methods
for various objects, generally providing terser output.  The
\code{\link{verbosity}} function can be used to make these
replace the standard print methods.
}
\seealso{
\code{\link{verbosity}}
}

