% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props-counts.R
\name{props}
\alias{props}
\alias{counts}
\title{Compute all proportions or counts}
\usage{
props(x, format = "proportion", ...)

counts(x, format = "count", ...)
}
\arguments{
\item{x}{an object}

\item{format}{a character string describing the desired format of the results.
One of \code{'default'}, \code{'count'}, \code{'proportion'}, \code{'percent'}, 
\code{'data.frame'}, \code{'sparse'}, or \code{'default'}.
In case of \code{'default'}, counts are used unless there is a condition, in
which case proportions are used instead.  Note that prior to version 0.9.3, 
\code{'default'} was the default, now it is \code{'count'}.
\code{'data.frame'} converts the table to a data frame with one row per cell;
\code{'sparse'} additionally removes any rows with 0 counts.}

\item{...}{additional arguments passed to \code{\link{table}}}
}
\description{
These wrappers around tally make it easy to compute proportions
and counts.
}
\examples{
props(HELPrct$substance)
props( ~ substance, data = HELPrct)
props( ~ substance | sex, data = HELPrct)
props( ~ substance | sex, data = HELPrct, format = "percent")
counts( ~ substance | sex, data = HELPrct)
if (require(ggformula)) {
  df_stats( ~ substance | sex, data = HELPrct, props, stats, format = "long")
}
}
\seealso{
\code{\link{prop}()}

\code{\link{count}()}
}
