% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{r.squared}
\alias{r.squared}
\title{Extract r-squared value}
\usage{
r.squared(x, ...)
}
\arguments{
\item{x}{an object}

\item{\dots}{additional arguments}
}
\description{
Attempts to extract an r-squared value from a model or model-like object.
}

