\docType{data}
\name{Ratings}
\alias{Ratings}
\title{Instructor Ratings}
\format{A data frame with 81 observations on the following variables.

   \itemize{
    \item{\code{score}}{score on the evaluation scale}
    \item{\code{situation}}{a factor with levels \code{Class I}, \code{Class II} or \code{Class III}}
  }}
\usage{
  data(Ratings)
}
\description{
  At the end of a statistics course, the 27 students in the
  class were asked to rate the instructor on a number scale
  of 1 to 9 (1 being "very poor", and 9 being "best
  instructor I've ever had"). This dataset contains three
  sets of hypothetical rating data.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

