% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic-package.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeFun}
\alias{fitdistr}
\alias{fractions}
\alias{lhs}
\alias{rhs}
\alias{condition}
\alias{counts}
\alias{props}
\alias{prop}
\alias{prop1}
\alias{perc}
\alias{percs}
\alias{count}
\alias{tally}
\alias{dfapply}
\alias{ediff}
\alias{inspect}
\alias{msummary}
\alias{n_missing}
\alias{logit}
\alias{ilogit}
\alias{compare}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{MASS}{\code{\link[MASS:fitdistr]{fitdistr()}}, \code{\link[MASS:fractions]{fractions()}}}

  \item{mosaicCore}{\code{\link[mosaicCore:compare]{compare()}}, \code{\link[mosaicCore:condition]{condition()}}, \code{\link[mosaicCore:count]{count()}}, \code{\link[mosaicCore:counts]{counts()}}, \code{\link[mosaicCore:counts]{counts()}}, \code{\link[mosaicCore:dfapply]{dfapply()}}, \code{\link[mosaicCore:ediff]{ediff()}}, \code{\link[mosaicCore:ilogit]{ilogit()}}, \code{\link[mosaicCore:inspect]{inspect()}}, \code{\link[mosaicCore:lhs]{lhs()}}, \code{\link[mosaicCore:logit]{logit()}}, \code{\link[mosaicCore:makeFun]{makeFun()}}, \code{\link[mosaicCore:makeFun]{makeFun()}}, \code{\link[mosaicCore:msummary]{msummary()}}, \code{\link[mosaicCore:n_missing]{n_missing()}}, \code{\link[mosaicCore:perc]{perc()}}, \code{\link[mosaicCore:percs]{percs()}}, \code{\link[mosaicCore:prop]{prop()}}, \code{\link[mosaicCore:prop1]{prop1()}}, \code{\link[mosaicCore:props]{props()}}, \code{\link[mosaicCore:rhs]{rhs()}}, \code{\link[mosaicCore:tally]{tally()}}}
}}

