% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{relm}
\alias{relm}
\title{Resample a Linear Model}
\usage{
relm(model, ..., envir = environment(formula(model)))
}
\arguments{
\item{model}{a linear model object produced using \code{\link{lm}}.}

\item{...}{additional arguments passed through to \code{\link{resample}}.}

\item{envir}{an environment in which to (re)evaluate the linear model.}
}
\description{
Residual resampling from a linear model
}
