% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rescale}
\alias{rescale}
\alias{rescale.character}
\alias{rescale.data.frame}
\alias{rescale.default}
\alias{rescale.factor}
\alias{rescale.numeric}
\title{Rescale}
\usage{
rescale(x, range, domain, ...)

\method{rescale}{data.frame}(x, range = c(0, 1), domain, ...)

\method{rescale}{factor}(x, range, domain = range(1:nlevels(x)), ...)

\method{rescale}{numeric}(x, range = c(0, 1), domain = range(x, na.rm =
  TRUE), ...)

\method{rescale}{default}(x, range = c(0, 1), domain, ...)

\method{rescale}{character}(x, range = c(0, 1), domain, ...)
}
\arguments{
\item{x}{an R object to rescale}

\item{domain}{a numeric vector of length 2}

\item{range}{a numeric vector of length 2}

\item{...}{additional arguments}
}
\description{
Rescale vectors or variables within data frames.  This can be useful
for comparing vectors that are on different scales, for example in
parallel plots or heatmaps.
}

