\name{rpoly2}
\alias{rpoly2}
\title{random 2nd degree polynomials}
\usage{
  rpoly2(seed = NULL)
}
\arguments{
  \item{seed}{random seed}
}
\value{
  a function defined by a 2nd degree polynomial of 2
  variables with coefficients selected randomly according
  to a Unif(-1,1) distribution.
}
\description{
  Generates a random 2nd degree poly of 2 vars (as a
  function)
}
\keyword{random}

