% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{rspin}
\alias{rspin}
\title{Simulate spinning a spinnner}
\usage{
rspin(n, probs, labels = 1:length(probs))
}
\arguments{
\item{n}{number of spins of spinner}

\item{probs}{a vector of probabilities.  If the sum is not 1, the 
probabiliies will be rescaled.}

\item{labels}{a character vector of labels for the categories}
}
\description{
This is essentially \code{rmultinom} with a different interface.
}
\examples{
rspin(20, prob=c(1,2,3), labels=c("Red", "Blue", "Green"))
do(2) * rspin(20, prob=c(1,2,3), labels=c("Red", "Blue", "Green"))
}

