% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{rsquared}
\alias{rsquared}
\title{Extract r-squared value}
\usage{
rsquared(x, ...)
}
\arguments{
\item{x}{an object}

\item{\dots}{additional arguments}
}
\description{
Attempts to extract an r-squared value from a model or model-like object.
}

