% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rstudio_is_available}
\alias{rstudio_is_available}
\title{Check whether RStudio is in use}
\usage{
rstudio_is_available()
}
\value{
a logical
}
\description{
This functions checks that RStudio is in use.  It will likely be removed
from this package once the versions of RStudio in popular use rely on the
manipulate package on CRAN which will provide its own version.
}

