% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select}
\alias{select}
\title{Select a subset of columns}
\usage{
select(.data, ...)
}
\description{
Select a subset of columns using \code{dplyr::select}.
}
\details{
Both the \pkg{MASS} and \pkg{dplyr} packages define a generic
\code{select}.  When the \pkg{mosaic} package is loaded,
\code{dplyr::select} will be preferred.
}
\examples{
HELPrct \%>\% sample(5) \%>\% select(sex, substance, homeless)
}
\seealso{
\code{\pkg{dplyr::}\link[dplyr]{select}}
and  \code{\pkg{MASS::}\link[MASS]{select}}
}

