\docType{data}
\name{Sleep}
\alias{Sleep}
\title{Sleeping Time}
\format{A data frame with 142 observations on the following variables.

   \itemize{
    \item{\code{sleep}} {sleeping time in a typical day (in hours)}
  }}
\usage{
  data(Sleep)
}
\description{
  As part of a large survey conducted at a large state
  university, a random sample of 142 students were asked:
  "How many hours do you sleep in a typical day?"
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

