\docType{data}
\name{Sleep2}
\alias{Sleep2}
\title{Sleeping Time of Undergraduate and Graduate Students}
\format{A data frame with 125 rows on the following variables.

   \itemize{
    \item{\code{hours}} {number of hours that a student sleeps in a typical day)}
    \item{\code{status}} {\code{Undergrad} or \code{Graduate} student}
  }}
\usage{
  data(Sleep2)
}
\description{
  A study was conducted at a large state university in
  order to compare the sleeping habits of undergraduate
  students to those of graduate students. Random samples of
  75 undergraduate students and 50 graduate students were
  chosen and each of the subjects was asked to report the
  number of hours he or she sleeps in a typical day. The
  hypothesis is that since undergraduate students are
  generally younger and party more during their years in
  school, they sleep less, on average, than graduate
  students.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

