% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2.R
\name{sp2df}
\alias{sp2df}
\title{Transforms a shapefile into a dataframe}
\usage{
sp2df(map, ...)
}
\arguments{
\item{map}{A map object of class \code{SpatialPolygonsDataFrame}}

\item{...}{Other arguments, currently ignored}
}
\value{
A dataframe, in which the first 7 columns hold geographical
information (ex: \code{long} and \code{lat})
}
\description{
This function takes in a shapefile (formal class of
\code{SpatialPolygonsDataFrame}) and transforms it into a dataframe
}
