% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardName}
\alias{standardName}
\title{Standardization of Geographic Names}
\usage{
standardName(x, standard)
}
\description{
Often different sources of geographical data will use different names for
the same region.  These utilities make it easier to merge data from different
sources by converting names to standardized forms.
}

