\name{summary.mosaic}
\alias{summary.mosaic}
\title{
Summarize a Mosaic Graph
}
\description{
A summary of a mosaic graph object includes the estimated area, numbers of vertices and edges, and counts of vertices grouped by the number of edges leading to them.
}
\usage{
summary.mosaic(m)
}
\arguments{
  \item{m}{
A mosaic object generated by the \link{mgraph} function.
}
}
\references{
Alroy, J. A simple graph theoretic method provides accurate range area estimates. To be submitted to \emph{Ecology Letters}.
}
\author{
John Alroy
}
\examples{
m <- mgraph(runif(20),runif(20))
summary(m)
}
