% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2.R
\name{theme_map}
\alias{theme_map}
\title{ggplot2 theme for maps}
\usage{
theme_map(base_size = 12)
}
\arguments{
\item{base_size}{the base font size for the theme.}
}
\description{
A very plain \pkg{ggplot2} theme that is good for maps.
}
\details{
This theme is largely based on an example posted by 
Winston Chang at the \pkg{ggplot2} Google group forum.
}

