\docType{data}
\name{Time}
\alias{Time}
\title{Exercise Habits of Male Students}
\format{A data frame with 20 observations on the following variables.

   \itemize{
    \item{\code{time1}} {time (in hours) spending on exercises in a typical week, sample 1 of size 20}
    \item{\code{time2}} {time (in hours) spending on exercises in a typical week, sample 2 of size 20}
    \item{\code{time3}} {time (in hours) spending on exercises in a typical week, sample 3 of size 20}
    \item{\code{time4}} {time (in hours) spending on exercises in a typical week, sample 4 of size 20}
  }}
\usage{
  data(Time)
}
\description{
  A health educator at a small college wants to determine
  whether the exercise habits of male students in the
  college are similar to the exercise habits of male
  college students in general. The educator chooses a
  random sample of 20 male students and records the time
  they spend exercising in a typical week.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

