\docType{data}
\name{TV1}
\alias{TV1}
\title{TV Time 1}
\format{A data frame with 40 rows on the following variables.

   \itemize{
    \item{\code{time}} {amount of time (in minutes) that an individual watched TV during a week}
    \item{\code{gender}} {\code{Female} or \code{Male}}
  }}
\usage{
  data(TV1)
}
\description{
  A researcher wanted to study whether or not men and women
  differ in the amount of time they watch TV during a week.
  A random sample of 40 adults was chosen (22 of whom were
  women and 18 of whom were men). At the end of the week,
  each of the 40 subjects reported the total amount of time
  (in minutes) that he or she watched TV during that week.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

