\docType{data}
\name{TV2}
\alias{TV2}
\title{TV Time 2}
\format{A data frame with 400 rows on the following variables.

   \itemize{
    \item{\code{time}} {amount of time (in minutes) that an individual watched TV during a week}
    \item{\code{gender}} {\code{Female} or \code{Male}}
  }}
\usage{
  data(TV2)
}
\description{
  A researcher wanted to study whether or not men and women
  differ in the amount of time they watch TV during a week.
  A random sample of 400 adults was chosen (191 of whom
  were women and 209 of whom were men). At the end of the
  week, each of the 400 subjects reported the total amount
  of time (in minutes) that he or she watched TV during
  that week.
}
\references{
  Part of the Carnegie Mellon University Online Learning
  Initiative datasets.
}
\keyword{datasets}

