\name{.check_for_quant_input}
\alias{.check_for_quant_input}
\title{The Data Distribution}
\usage{
  .check_for_quant_input(x)
}
\arguments{
  \item{x}{a vector}
}
\value{
  TRUE if quantitative; else \code{stop}
}
\description{
  Density, distribution function, quantile function, and
  random generation from data.
}
\keyword{Check}
\keyword{distribution}
\keyword{input}
\keyword{is}
\keyword{quantitative}
\keyword{whether}

