\name{undocumented}
\alias{dotPlot}
\alias{panel.dotPlot}
\alias{surround}
\alias{adapt_seq}
\alias{cps}
\alias{galton}
\alias{gestation}
\alias{kidsfeet}
\alias{marriage}
\alias{sat}
\alias{utilities}
\alias{whickham}

\title{Undocumented Objects}

\description{Documentation is still pending for these objects}

\author{Randall Pruim}
