% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{vector2df}
\alias{vector2df}
\title{Convert a vector to a data frame}
\usage{
vector2df(x, nice_names = FALSE)
}
\arguments{
\item{x}{a vector}

\item{nice_names}{a logical indicating whether names should be nicified}
}
\value{
a data frame
}
\description{
Convert a vector into a 1-raw data frame using the names of the vector as
column names for the data frame
}
\examples{
vector2df(c(1, b = 2, `(Intercept)` = 3))
vector2df(c(1, b = 2, `(Intercept)` = 3), nice_names = TRUE)

}

