% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatingMeta.R
\name{aggregatingFunction1}
\alias{aggregatingFunction1}
\title{1-ary Aggregating functions}
\usage{
aggregatingFunction1(fun, output.multiple = FALSE, envir = parent.frame(),
  na.rm = getOption("na.rm", FALSE), style = c("flexible", "formula"))
}
\arguments{
\item{fun}{a function that takes a numeric vector and computes a summary statistic,
returning a numeric vector of length 1.}

\item{output.multiple}{a boolean indicating whether \code{fun} returns multiple values}

\item{envir}{an environment in which evaluation takes place.}

\item{na.rm}{the default value for na.rm in the resulting function.}

\item{style}{one of \code{"flexible"} or \code{"formula"}.  In the latter case, the first
argument must be a formula or evaluate to an object.  In the former case, bare names will
be converted into formulas.}
}
\value{
a function that generalizes \code{fun} to handle a formula/data frame interface.
}
\description{
\code{aggregatinFuntion1} creates statistical summaries of one numerical vector that are formula aware.
}
\examples{
if (require(mosaicData)) {
  foo <- aggregatingFunction1( base::mean )
  foo( ~length, data=KidsFeet )
  base::mean(KidsFeet$length)
  foo( length ~ sex, data=KidsFeet )
} 
}

