% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cross}
\alias{cross}
\title{Factor cross products}
\usage{
cross(..., sep = ":", drop.unused.levels = FALSE)
}
\arguments{
\item{sep}{separator between levels}

\item{drop.unused.levels}{should levels that do not appear in cross product be dropped?}

\item{\dots}{factors to be crossed.}
}
\value{
a factor
}
\description{
Construct a product of factors.
}
\examples{
x <- letters[1:3]
y <- c(1,2,1,1,3,1,3)
cross(x, y)
cross(x, y, drop.unused.levels=TRUE)
}
\keyword{manipulate}

