% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R, R/fetchGapminder.R, R/fetchGoogle.R
\name{fetchData}
\alias{fetchData}
\alias{fetchGapminder}
\alias{fetchGapminder1}
\alias{fetchGoogle}
\title{Defunct functions now in the fetch package}
\usage{
fetchData(...)

fetchGapminder1(...)

fetchGapminder(...)

fetchGoogle(...)
}
\arguments{
\item{...}{arguments}
}
\description{
These functions have been moved to the fetch package.
}

