% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfapply.R
\name{dfapply}
\alias{dfapply}
\title{apply-type function for data frames}
\usage{
dfapply(data, FUN, select = TRUE, ...)
}
\arguments{
\item{data}{data frame}

\item{FUN}{a function to apply to (some) variables in the data frame}

\item{select}{function used to select variables to which \code{FUN} is applied.  See examples.}

\item{\dots}{arguments passed along to \code{FUN}}
}
\description{
An \code{apply}-type function for data frames.
}
\examples{
dfapply(iris, favstats)
if (require(mosaicData)) {
  dfapply(HELPrct, table, select=is.factor)
  do.call(rbind, dfapply(HELPrct, favstats, select=is.numeric))
}
}
\seealso{
\code{\link{apply}},
\code{\link{sapply}},
\code{\link{tapply}},
\code{\link{lapply}}
}

