% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diffmean}
\alias{diffmean}
\alias{diffprop}
\title{Difference in means and proportions}
\usage{
diffmean(...)

diffprop(...)
}
\arguments{
\item{...}{arguments passed to \code{mean}}
}
\description{
Wrappers around \code{diff(mean(...))} and \code{diff(prop(...))} that
facilitate better naming of the result
}
\examples{
if (require(mosaicData)) {
diffprop( homeless ~ sex , data=HELPrct)
do(3) * diffprop( homeless ~ shuffle(sex) , data=HELPrct)
diffmean( age ~ substance, data=HELPrct)
do(3) * diffmean(age ~ shuffle(substance), data=HELPrct)
diffmean( age ~ sex, data=HELPrct)
do(3) * diffmean(age ~ shuffle(sex), data=HELPrct)
}
}

