% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docFile}
\alias{docFile}
\title{Return the path to a documentation file in a package}
\usage{
docFile(file, package = "mosaic", character.only = FALSE)
}
\arguments{
\item{file}{the name of a file}

\item{package}{the name of a package}

\item{character.only}{a logical. If \code{TRUE} package names must be
specified as character, else names will be converted as a convenience
as is \code{\link{library}} and \code{\link{library}}.}
}
\value{
a character vector specifying the path to the file on the user's system.
}
\description{
Return the path to a documentation file in a package
}
\examples{
MustangPrice <- read.file(docFile("MustangPrice.csv", "mosaic"))
}

