% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fortify.summary.lm}
\alias{fortify.TukeyHSD}
\alias{fortify.summary.lm}
\title{Extract data from R objects}
\usage{
\method{fortify}{summary.lm}(model, data = NULL, level = 0.95, ...)

\method{fortify}{TukeyHSD}(model, data, ...)
}
\arguments{
\item{level}{confidence level}

\item{model}{an R object}

\item{data}{original data set, if needed}

\item{...}{additional arguments}
}
\description{
Extract data from R objects
}
\examples{
fortify(TukeyHSD(lm(age ~ substance, data=HELPrct)))
}

