% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalFormula.R
\name{joinFrames}
\alias{joinFrames}
\alias{joinTwoFrames}
\title{Join data frames}
\usage{
joinFrames(...)

joinTwoFrames(left, right)
}
\arguments{
\item{\dots}{data frames to be joined}

\item{left, right}{data frames}
}
\value{
a data frame containing columns from each of data frames being joined.
}
\description{
Join data frames
}
