% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MAD}
\alias{MAD}
\alias{SAD}
\title{All pairs mean and sum of absolute differences}
\usage{
MAD(x, ..., na.rm = getOption("na.omit", FALSE))

SAD(x, ..., na.rm = getOption("na.omit", FALSE))
}
\arguments{
\item{x}{a numeric vector}

\item{...}{if present, appended to x}

\item{na.rm}{a logical indicating whether NAs should be removed before
calculaing.}
}
\value{
the mean or sum of the absolute differences between each pair
of values in \code{c(x,...)}.
}
\description{
All pairs mean and sum of absolute differences
}
\examples{
SAD(1:3)
MAD(1:3)
}
\seealso{
\code{link{mad}}
}

