% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MAD_}
\alias{MAD_}
\alias{SAD_}
\title{All pairs mean and sum of absolute differences}
\usage{
MAD_(x, ..., na.rm = getOption("na.omit", FALSE))

SAD_(x, ..., na.rm = getOption("na.omit", FALSE))
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{if present, appended to x}

\item{na.rm}{a logical indicating whether NAs should be removed before
calculaing.
#'}

\item{..fun..}{the underlyin function used in the computation}

\item{groups}{a grouping variable, typically a name of a variable in \code{data}}

\item{data}{a data frame in which to evaluate formulas (or bare names).
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were
a data frame.  But this may not be appropriate for programming uses.
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned.}
}
\value{
the mean or sum of the absolute differences between each pair
of values in \code{c(x,...)}.
}
\description{
All pairs mean and sum of absolute differences
}
\seealso{
\code{link{mad}}
}

