% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatingMeta.R
\name{MAD_}
\alias{MAD_}
\alias{SAD_}
\title{All pairs mean and sum of absolute differences}
\usage{
MAD_(x, ..., na.rm = getOption("na.omit", FALSE))

SAD_(x, ..., na.rm = getOption("na.omit", FALSE))
}
\arguments{
\item{x}{a numeric vector or a formula.}

\item{...}{additional arguments appended to \code{x}}

\item{na.rm}{a logical indicating whether NAs should be removed before
calculaing.}
}
\value{
the mean or sum of the absolute differences between each pair
of values in \code{c(x,...)}.
}
\description{
All pairs mean and sum of absolute differences
}
\seealso{
\code{\link{mad}}
}

