% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mid}
\alias{mid}
\title{midpoints along a sequence}
\usage{
mid(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of length 1 less than \code{x}
}
\description{
Compute a vector of midpoints between values in a numeric vector
}
\examples{
mid(1:5)
mid((1:5)^2)
}

