% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatingAux.R, R/do.R, R/utils.R
\name{.is.formula}
\alias{.is.formula}
\alias{.is.simple.formula}
\alias{.simple.part}
\alias{.make.data.frame}
\alias{.merge_data_frames}
\alias{.squash_names}
\alias{.do.safe.call}
\title{Check if formula}
\usage{
.is.formula(x)

.is.simple.formula(x)

.simple.part(x)

.make.data.frame(x)

.merge_data_frames(a, b)

.squash_names(object, sep = ":")

.do.safe.call(what, args, quote = FALSE, envir = parent.frame(), ...)
}
\arguments{
\item{x}{an object}

\item{a}{a data frame}

\item{b}{a data frame}

\item{object}{an object}

\item{sep}{a character}

\item{what}{either a function or a non-empty character string naming the function to be called.}

\item{args}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{quote}{a logical value indicating whether to quote the arguments.}

\item{envir}{an environment within which to evaluate the call.
This will be most useful if what is a character string and the arguments are symbols or quoted expressions.}

\item{x}{a formula}

\item{x}{a formula}

\item{x}{object to be converted}
}
\value{
TRUE for a formula, FALSE otherwise, even if evaluation throws an error

TRUE if formula has no left-hand side or a simple right-hand side 
(e.g., \code{NULL}, ., 1,  or 0)

simple part of formula or NULL if formula is not simple

a data frame

a data frame

a character vector

The result of the (evaluated) function call.
}
\description{
Check if formula

Check for simple formula

Extract simple part from formula
}
\details{
\code{.make.data.frame} converts things to a data frame

\code{.merge_data_frames} is a wrapper around merge

\code{.squash_names} squashes names of a data frame into a single string

\code{.do.safe.call} avoids conflicts between named arguments and ... by taking named arguments
preferentially.
}
\seealso{
\code{\link{do.call}}
}
\keyword{internal}
