\docType{package}
\name{mosaic-package}
\alias{mosaic}
\alias{mosaic-package}
\title{mosaic: the Project MOSAIC package}
\description{
  mosaic
}
\details{
  Data sets and utilities from Project MOSAIC
  (mosaic-web.org) used to teach mathematics, statistics,
  computation and modeling.  Funded by the NSF, Project
  MOSAIC is creating a community of educators working to
  tie together aspects of quantitative work that students
  in science, technology, engineering and mathematics will
  need in their professional lives, but which are usually
  taught in isolation, if at all.

  \tabular{ll}{ Package: \tab mosaic\cr Type: \tab
  Package\cr Version: \tab 0.3-2\cr Date: \tab
  2012-01-12\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Randall Pruim (\email{rpruim@calvin.edu}), Daniel Kaplan
  (\email{kaplan@macalester.edu})
}
\references{
  \url{http://mosaic-web.org}
}
\keyword{package}

