% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mosaic-package}
\alias{mosaic}
\alias{mosaic-package}
\title{mosaic: the Project MOSAIC package}
\description{
mosaic
}
\details{
Data sets and utilities from Project MOSAIC (mosaic-web.org) used to teach mathematics,
statistics, computation and modeling.  Funded by the NSF, Project MOSAIC is a
community of educators working to tie together aspects of quantitative work that students
in science, technology, engineering and mathematics will need in their professional lives,
but which are usually taught in isolation, if at all.
}
\author{
Randall Pruim (\email{rpruim@calvin.edu}), Daniel Kaplan (\email{kaplan@macalester.edu}), Nicholas Horton (\email{nhorton@smith.edu})
}
\references{
\url{http://mosaic-web.org}
}
\keyword{package}

