% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mplot}
\alias{mplot}
\alias{mplot.TukeyHSD}
\alias{mplot.data.frame}
\alias{mplot.lm}
\alias{mplot.summary.lm}
\title{Generic plotting}
\usage{
mplot(object, ...)

\method{mplot}{lm}(object, which = c(1:3, 7), system = c("lattice",
  "ggplot2", "base"), ask = FALSE, multiplot = "package:gridExtra" \%in\%
  search(), par.settings = theme.mosaic(), level = 0.95,
  title = paste("model: ", deparse(object$call), "\\n"), ...)

\method{mplot}{data.frame}(object, default = plotTypes,
  system = c("lattice", "ggplot2"), show = FALSE, title = "", ...)

\method{mplot}{summary.lm}(object, system = c("lattice", "ggplot2"),
  level = 0.95, par.settings = trellis.par.get(), ...)

\method{mplot}{TukeyHSD}(object, system = c("lattice", "ggplot2"),
  ylab = "", xlab = "difference in means",
  title = "Tukey's Honest Significant Differences",
  par.settings = trellis.par.get(), ...)

\method{mplot}{TukeyHSD}(object, system = c("lattice", "ggplot2"),
  ylab = "", xlab = "difference in means",
  title = "Tukey's Honest Significant Differences",
  par.settings = trellis.par.get(), ...)
}
\arguments{
\item{object}{an R object from which a plot will be constructed.}

\item{data}{a data frame containing the variables that might be used in the plot.
Note that for maps, the data frame must contain coordinates of the polygons
comprising the map and a variable for determining which corodiantes are part
of the same region.  See \code{\link{sp2df}} for one way to create such
a data frame.  Typically \code{\link{merge}} will be used to combine the map
data with some auxilliary data to be displayed as fill color on the map, although
this is not necessary if all one wants is a map.}

\item{default}{default type of plot to create; one of
\code{"scatter"},
\code{"jitter"},
\code{"boxplot"},
\code{"violin"},
\code{"histogram"},
\code{"density"},
\code{"frequency polygon"},
\code{"xyplot"},
or
\code{"map"}.  Unique prefixes suffice.}

\item{system}{which graphics system to use (initially) for plotting (\pkg{ggplot2}
or \pkg{lattice}).  A check box will allow on the fly change of plotting system.}

\item{show}{a logical, if \code{TRUE}, the code will be displayed each time the plot is
changed.}

\item{which}{a numeric vector used to select from 7 potential plots}

\item{ask}{if TRUE, each plot will be displayed separately after the user
responds to a prompt.}

\item{multiplot}{if TRUE and \code{ask == FALSE}, all plots will be
displayed together.}

\item{title}{title for plot}

\item{...}{additional arguments.  If \code{object} is an \code{lm}, subsets
of these arguments are passed to \code{grid.arrange} and to the
\pkg{lattice} plotting routines; in particular,
\code{nrow} and \code{ncol} can be used to control the number of rows
and columns used.}

\item{level}{a confidence level}

\item{par.settings}{\pkg{lattice} theme settings}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}
}
\value{
Nothing.  Just for side effects.
}
\description{
Generic function plotting for R objects.  Currently plots exist for
\code{data.frame}s, \code{lm}s, (including \code{glm}s).
}
\examples{
mplot( lm( width ~ length + sex, data=KidsFeet) )
\dontrun{
mplot( HELPrct )
mplot( HELPrct, "histogram" )
}
mplot(summary(lm(width ~ length * sex, data=KidsFeet)))
mplot(TukeyHSD( lm(age ~ substance, data=HELPrct) ) )
mplot(TukeyHSD( lm(age ~ substance, data=HELPrct) ), system="ggplot2" )
mplot(TukeyHSD( lm(age ~ substance, data=HELPrct) ) )
mplot(TukeyHSD( lm(age ~ substance, data=HELPrct) ), system="ggplot2" )
}

