% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwiseModels.R
\name{MSPE}
\alias{MSPE}
\title{Mean Squared Prediction Error}
\usage{
MSPE(model, data)
}
\arguments{
\item{model}{a model produced by \code{lm}, \code{glm}, or \code{gwm}.}

\item{data}{a data frame}
}
\description{
Mean Squared Prediction Error
}
\examples{
HELP <- HELPrct \%>\% sample_frac(.3)
MSPE( gwm( age ~ sex, data = HELP), HELPrct)
MSPE( gwm( age ~ 1, data = HELP), HELPrct)
MSPE( gwm( age ~ sex + homeless, data = HELP), HELPrct)
MSPE( gwm( sex ~ 1, data = HELP), HELPrct)
MSPE( gwm( sex ~ homeless, data = HELP), HELPrct)
MSPE( gwm( sex ~ homeless + substance, data = HELP), HELPrct)
}

