% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.R
\name{nice_names}
\alias{nice_names}
\title{Nice names}
\usage{
nice_names(x, unique = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{unique}{a logical indicating whether returned values should be uniquified.}
}
\value{
a character vector
}
\description{
Convert a character vector into a similar character vector that would
work better as names in a data frame by avoiding certain awkward characters
}
\examples{
nice_names( c("bad name", "name (crazy)", "a:b", "two-way") )
}

