% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ntiles}
\alias{ntiles}
\title{Create vector based on roughly equally sized groups}
\usage{
ntiles(x, n = 3, format = c("rank", "interval", "center", "left", "right"))
}
\arguments{
  \item{x}{a numeric vector}

  \item{n}{number of quantiles}

  \item{format}{a specification of desired output format.}
}
\value{
a vector.  The type of vector will depend on \code{format}.
}
\description{
Create vector based on roughly equally sized groups
}
\examples{
ntiles(1:50, 4)
ntiles(1:50, 4, format="center")
ntiles(1:50, 4, format="interval")
ntiles(1:50, 4, format="left")
ntiles(1:50, 4, format="right")
}

