% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFun.R
\name{panel.levelcontourplot}
\alias{panel.levelcontourplot}
\title{Lattice plot that draws a filled contour plot}
\usage{
panel.levelcontourplot(x, y, z, subscripts = 1, at, shrink, labels = TRUE,
  label.style = c("mixed", "flat", "align"), contour = FALSE,
  region = TRUE, col = add.line$col, lty = add.line$lty,
  lwd = add.line$lwd, border = "transparent", ...,
  col.regions = regions$col, filled = TRUE, alpha.regions = regions$alpha)
}
\arguments{
\item{x}{x on a grid}

\item{y}{y on a grid}

\item{z}{zvalues for the x and y}

\item{subscripts}{which points to plot}

\item{at}{cuts for the contours}

\item{shrink}{what does this do?}

\item{labels}{draw the contour labels}

\item{label.style}{where to put the labels}

\item{contour}{logical draw the contours}

\item{region}{logical color the regions}

\item{col}{color for contours}

\item{lty}{type for contours}

\item{lwd}{width for contour}

\item{border}{type of border}

\item{...}{dots additional arguments}

\item{col.regions}{a vector of colors or a function (\code{topo.colors} by default) for generating such}

\item{filled}{whether to fill the contours with color}

\item{alpha.regions}{transparency of regions}
}
\description{
Used within plotFun
}

