% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdist}
\alias{pdist}
\title{Illustrated probability calculations from distributions}
\usage{
pdist(dist = "norm", q, plot = TRUE, verbose = FALSE, invisible = FALSE,
  digits = 4, lower.tail = TRUE, log.p = FALSE, xlim, ylim, vlwd = 2,
  vcol = trellis.par.get("add.line")$col, rot = 45, ...)
}
\arguments{
\item{dist}{a character discription of a distribution, for example
\code{"norm"}, \code{"t"}, or \code{"chisq"}}

\item{q}{a vector of quantiles}

\item{plot}{a logical indicating whether a plot should be created}

\item{verbose}{a logical}

\item{invisible}{a logical}

\item{digits}{the number of digits desired}

\item{lower.tail}{a logical indicating whether lower tail probabilities are used}

\item{log.p}{a logical indicating whether the log of the probability should be
calculated}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{vlwd}{width of vertical lines}

\item{vcol}{color of vertical lines}

\item{rot}{angle for rotating text indicating probability}

\item{...}{additional arguments, including parameters of the distribution
and additional options for the plot}
}
\value{
a vector of quantiles; often used for side effect of creating a plot
}
\description{
Illustrated probability calculations from distributions
}
\examples{
pdist("norm", -2:2)
pdist("norm", seq(80,120, by=10), mean=100, sd=10)
pdist("chisq", 2:4, df=3)
pdist("f", 1, df1=2, df2=10)
}

