% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic-package.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeFun}
\alias{reexports}
\alias{fitdistr}
\alias{reexports}
\alias{fractions}
\alias{reexports}
\alias{lhs}
\alias{reexports}
\alias{rhs}
\alias{reexports}
\alias{condition}
\alias{reexports}
\alias{makeFun}
\alias{reexports}
\alias{counts}
\alias{reexports}
\alias{props}
\alias{reexports}
\alias{prop}
\alias{reexports}
\alias{prop1}
\alias{reexports}
\alias{perc}
\alias{reexports}
\alias{count}
\alias{reexports}
\alias{tally}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{MASS}{\code{\link[MASS]{fitdistr}}, \code{\link[MASS]{fractions}}}

  \item{mosaicCore}{\code{\link[mosaicCore]{makeFun}}, \code{\link[mosaicCore]{lhs}}, \code{\link[mosaicCore]{rhs}}, \code{\link[mosaicCore]{condition}}, \code{\link[mosaicCore]{makeFun}}, \code{\link[mosaicCore]{counts}}, \code{\link[mosaicCore]{props}}, \code{\link[mosaicCore]{prop}}, \code{\link[mosaicCore]{prop1}}, \code{\link[mosaicCore]{perc}}, \code{\link[mosaicCore]{count}}, \code{\link[mosaicCore]{tally}}}
}}

