% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{repeater-class}
\alias{repeater-class}
\title{Repeater objects}
\description{
Repeater objects can be used with the \code{*} operator to repeat
things multiple time using a different syntax and different output
format from that used by, for example, \code{\link{replicate}}.
}
\section{Slots}{
  \describe{ \item{\code{n}:}{Object of class
  \code{"numeric"} indicating how many times to repeat
  something.} \item{\code{cull}:}{Object of class
  \code{"function"} that culls the ouput from each
  repetition.} \item{\code{mode}:}{Object of class
  \code{"character"} indicating the output mode ('default',
  'data.frame', 'matrix', 'vector', or 'list').  For most
  purposes 'default' (the default) should suffice.}
  \item{\code{algorithm}:}{an algorithm number.} }
}
\seealso{
\code{\link{do}}
}

