\name{repeater-class}
\alias{repeater-class}
\title{Repeater objects}
\description{
  Repeater objects can be used with the \code{*} operator
  to repeat things multiple time using a different syntax
  and different output format from that used by, for
  example, \code{\link{replicate}}.
}
\details{
  The class is \itemize{ \item{\code{n}} number of times to
  repeat something \item{\code{cull}} a function used to
  cull output \item{\code{mode}} the mode used for results
  (NULL or data frame or matrix) }
}
\seealso{
  \code{\link{do}}
}

