% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\alias{rescale.data.frame}
\alias{rescale.factor}
\alias{rescale.numeric}
\alias{rescale.default}
\alias{rescale.character}
\title{Rescale}
\usage{
rescale(x, range, domain = NULL, ...)

\method{rescale}{data.frame}(x, range = c(0, 1), domain = NULL, ...)

\method{rescale}{factor}(x, range, domain = NULL, ...)

\method{rescale}{numeric}(x, range = c(0, 1), domain = NULL, ...)

\method{rescale}{default}(x, range = c(0, 1), domain = NULL, ...)

\method{rescale}{character}(x, range = c(0, 1), domain = NULL, ...)
}
\arguments{
\item{x}{an R object to rescale}

\item{range}{a numeric vector of length 2}

\item{domain}{a numeric vector of length 2 or \code{NULL}}

\item{...}{additional arguments}
}
\description{
Rescale vectors or variables within data frames.  This can be useful
for comparing vectors that are on different scales, for example in
parallel plots or heatmaps.
}
