% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardName}
\alias{standardCountry}
\alias{standardName}
\alias{standardState}
\title{Standardization of Geographic Names}
\usage{
standardName(x, standard)

standardCountry(x)

standardState(x)
}
\arguments{
\item{x}{A vector with the names to standardize}

\item{standard}{A named vector in which each name is a particular spelling of
the name in question and the value is the standardized version of that name}

\item{x}{A vector with the country names to standardize}

\item{x}{A vector with the US state names to standardize}
}
\value{
A vector with standardized country names (in the ISO-a3 format)

A vector with standardized state names (as two-letter abbreviations)
}
\description{
Often different sources of geographical data will use different names for
the same region.  These utilities make it easier to merge data from different
sources by converting names to standardized forms.
}

