\name{undocumented}
\alias{dotPlot}
\alias{panel.dotPlot}
\alias{surround}
\alias{adapt_seq}
\alias{cps}
\alias{galton}
\alias{gestation}
\alias{kidsfeet}
\alias{marriage}
\alias{whickham}
\alias{latlon2xyz}
\alias{xyz2latlon}
\alias{rad2deg}
\alias{deg2rad}
\alias{statTally}

\title{Undocumented Objects}

\description{Documentation is still pending for these objects}

\author{Randall Pruim}
