% Generated by roxygen2 (4.0.0): do not edit by hand
\name{vlength}
\alias{vlength}
\title{Vector Length}
\usage{
vlength(x, ...)
}
\arguments{
  \item{x}{a vector}

  \item{\dots}{ignored for now}
}
\value{
the length of the vector (i.e., the square root of the sum of the squares of the components)
}
\description{
Compute the length of a numeric vector in the usual mathematica sense of vector length
}
\examples{
vlength(rep(1,4))
m <- lm( length ~ width, data=KidsFeet )
# These should be the same
vlength( m$effects )
vlength( KidsFeet$length)
# So should these
vlength( tail(m$effects, -2) )
sqrt(sum(resid(m)^2))
}
\seealso{
\code{link{project}}
}

