% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xchisq.test}
\alias{xchisq.test}
\title{Augmented Chi-squared test}
\usage{
xchisq.test(x, y = NULL, correct = TRUE, p = rep(1/length(x), length(x)),
  rescale.p = FALSE, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{x,y,correct,p,rescale.p,simulate.p.value}{}

\item{correct}{A logical indicating whether the Yates' continuity correction should be applied
to 2 by 2 tables (when not simulating).}

\item{\dots}{Arguments passed directly to \code{\link{chisq.test}}.}

\item{simulate}{A logical indicating whether the p-value should be computed by simulation.}
}
\description{
This augmented version of \code{\link{chisq.test}} provides more verbose
output.
}
\examples{
# Physicians' Health Study data
phs <- cbind(c(104,189),c(10933,10845))
rownames(phs) <- c("aspirin","placebo")
colnames(phs) <- c("heart attack","no heart attack")
phs
xchisq.test(phs)
}
\seealso{
\code{\link{chisq.test}}
}

